Move isIE11 module
